requirements.txt
streamlit
tensorflow==2.10.0
numpy
Pillow